Implemented Service Locator with Dependency Injection.
- Implement a Service Locator combined with Dependency Injection to manage shared services.
- Refactor Manager scripts into Services.
- Apply proper namespaces in all scripts.
