+++ Assets/Prefabs/GameScreenCanvas.prefab
@@ -2443,7 +2443,7 @@ MonoBehaviour:
   m_HandleRect: {fileID: 959120782583649624}
   m_Direction: 0
   m_Value: 0
-  m_Size: 0.48914433
+  m_Size: 1
   m_NumberOfSteps: 0
   m_OnValueChanged:
     m_PersistentCalls:
@@ -2854,6 +2854,7 @@ GameObject:
   - component: {fileID: 8604574825388780430}
   - component: {fileID: 2352251241997428113}
   - component: {fileID: 8481916440574053956}
+  - component: {fileID: 861794354411791806}
   m_Layer: 5
   m_Name: GameScreenCanvas
   m_TagString: Untagged
@@ -2947,6 +2948,31 @@ MonoBehaviour:
   m_BlockingMask:
     serializedVersion: 2
     m_Bits: 4294967295
+--- !u!114 &861794354411791806
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6832869865635585097}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f0bc2fff94e17f841951ed1e3168162c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  currencyPrefab: {fileID: 7852301056319017168, guid: 115a7cfb2e7868944995755752eed27e, type: 3}
+  chestPrefab: {fileID: 1442585979719524587, guid: 609e66cd92c6b08419834ee1e992ab92, type: 3}
+  generateChestButton: {fileID: 4290044072070741704}
+  currencyPanel: {fileID: 649832965242984896}
+  chestSlotContentPanel: {fileID: 5096889197095828938}
+  chestProcessingPanel: {fileID: 6470075613631912754}
+  chestProcessingActionOneButton: {fileID: 4411592685450985852}
+  chestProcessingActionOneText: {fileID: 275026012386778629}
+  chestProcessingActionTwoButton: {fileID: 4190330223323271850}
+  chestProcessingActionTwoText: {fileID: 3442347083096793813}
+  chestProcessingCloseButton: {fileID: 1555316164235820799}
+  notificationPopupPanel: {fileID: 8024265590500537673}
+  notificationPopupText: {fileID: 2273575857109194722}
 --- !u!1 &7106033575544365925
 GameObject:
   m_ObjectHideFlags: 0
@@ -3509,7 +3535,7 @@ GameObject:
   - component: {fileID: 3772202170635040015}
   - component: {fileID: 650158058814737607}
   m_Layer: 5
-  m_Name: MetricsPanel
+  m_Name: CurrencyPanel
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
