+++ Assets/Scriptable Objects/ChestConfig.asset
@@ -12,6 +12,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 6fbac56005fb4864ca2f53b8f8485d93, type: 3}
   m_Name: ChestConfig
   m_EditorClassIdentifier: 
+  chestPrefab: {fileID: 1442585979719524587, guid: 609e66cd92c6b08419834ee1e992ab92, type: 3}
   minChestCount: 4
   maxChestCount: 10
   chests:
