+++ Assets/Scriptable Objects/CurrencyConfig.asset
@@ -18,6 +18,6 @@ MonoBehaviour:
     currencyImage: {fileID: 10913, guid: 0000000000000000f000000000000000, type: 0}
     imageColor: {r: 0.8773585, g: 0.8559332, b: 0.25161976, a: 0}
   - currencyType: 1
-    initialValue: 99
+    initialValue: 999
     currencyImage: {fileID: 10905, guid: 0000000000000000f000000000000000, type: 0}
     imageColor: {r: 0.115454145, g: 0.3962264, b: 0.061302926, a: 0}
