+++ Assets/Scripts/Chest/ChestService.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c0da11e2ad59e6b4a8e06394dfd0be50
+guid: eece52bbb47999e48bd84937a2bd3268
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
