+++ Assets/Scripts/Chest/ChestType.cs
@@ -1,16 +1,19 @@
-public enum ChestType
+namespace ServiceLocator.Chest
 {
-    Common,
-    Rare,
-    Epic,
-    Legendary
-}
+    public enum ChestType
+    {
+        Common,
+        Rare,
+        Epic,
+        Legendary
+    }
 
-public enum ChestState
-{
-    Locked,
-    Unlock_Queue,
-    Unlocking,
-    Unlocked,
-    Collected
+    public enum ChestState
+    {
+        Locked,
+        Unlock_Queue,
+        Unlocking,
+        Unlocked,
+        Collected
+    }
 }
