+++ Assets/Scripts/Currency/CurrencyConfig.cs
@@ -1,16 +1,19 @@
 using UnityEngine;
 
-[CreateAssetMenu(fileName = "CurrencyConfig", menuName = "ScriptableObjects/CurrencyConfig")]
-public class CurrencyConfig : ScriptableObject
+namespace ServiceLocator.Currency
 {
-    public CurrencyData[] currencies;
-}
+    [CreateAssetMenu(fileName = "CurrencyConfig", menuName = "ScriptableObjects/CurrencyConfig")]
+    public class CurrencyConfig : ScriptableObject
+    {
+        public CurrencyData[] currencies;
+    }
 
-[System.Serializable]
-public class CurrencyData
-{
-    public CurrencyType currencyType;
-    public int initialValue;
-    public Sprite currencyImage;
-    public Color imageColor;
+    [System.Serializable]
+    public class CurrencyData
+    {
+        public CurrencyType currencyType;
+        public int initialValue;
+        public Sprite currencyImage;
+        public Color imageColor;
+    }
 }
