+++ Assets/Scripts/Currency/CurrencyService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 59101dfb56bdd1144bb42d6a12511d04
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
