+++ Assets/Scripts/Main/GameService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 17508054a2c75e840bc18eb2cae66a7f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
