+++ Assets/Scripts/UI/UIService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f0bc2fff94e17f841951ed1e3168162c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
